requirements.txt
langchain-mongodb
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
pymongo
python-dotenv
apache-airflow 
pypdf
langchain-groq